🔧 chore : 전역 메타데이터 설정
